requirements.txt
flask==2.0.1
praw==7.7.1
python-dotenv==0.19.2
torch==2.0.1
transformers==4.31.0
accelerate==0.21.0
bitsandbytes==0.41.1
pillow==9.5.0
requests==2.31.0
tenacity==8.2.2
sentencepiece==0.1.99
protobuf==4.25.1
loguru==0.7.0